machine/riscv: simplify overriding refactor done in #1090 and #1092
While adding more downstream changes, I realized the number of rv_mem*.h
etc headers can quickly grow as we specialize more and more APIs from
'string.h' etc. Put all the related changes into one header, e.g.
'rv_string.h' for all 'string.h' related API specializations, to avoid
the potential header file clutter.

The existing 'rv_string.h' is renamed to 'rv_strcpy.h'.
